refactor(SCIM): update imports after SSO co-location
Follow-up to the rename commit — point all imports at the new
`tabs/sso/...` paths and convert previously-relative imports
(`./base/forms/Button`, `./icons/Icon`, etc.) to absolute project
paths now that the depth changed.
