+++ frontend/web/components/pages/organisation-settings/OrganisationSettingsPage.tsx
@@ -13,7 +13,7 @@ import { LicensingTab } from './tabs/LicensingTab'
 import { CustomFieldsTab } from './tabs/CustomFieldsTab'
 import { APIKeysTab } from './tabs/APIKeysTab'
 import { WebhooksTab } from './tabs/WebhooksTab'
-import { SSOTab } from './tabs/SSOTab'
+import { SSOTab } from './tabs/sso/SSOTab'
 
 type OrganisationSettingsTab = {
   component: ReactNode
