+++ frontend/web/components/pages/organisation-settings/tabs/sso/saml/modals/CreateSAML.tsx
@@ -20,7 +20,7 @@ import { cloudDownloadOutline } from 'ionicons/icons'
 import Tabs from 'components/navigation/TabMenu/Tabs'
 import TabItem from 'components/navigation/TabMenu/TabItem'
 import { AttributeName } from 'common/types/responses'
-import SAMLAttributeMappingTable from 'components/SAMLAttributeMappingTable'
+import SAMLAttributeMappingTable from 'components/pages/organisation-settings/tabs/sso/saml/SAMLAttributeMappingTable'
 import Input from 'components/base/forms/Input'
 import Icon from 'components/icons/Icon'
 import Project from 'common/project'
